change banner notification location
